+++ src/js/loader.rs
@@ -3,7 +3,7 @@
 use crate::js::constant::WINDOWS_REGEX;
 use crate::js::module::ModulePath;
 use crate::js::module::ModuleSource;
-// use crate::js::module::CORE_MODULES;
+use crate::js::module::CORE_MODULES;
 use crate::js::transpiler::TypeScript;
 use crate::js::transpiler::Wasm;
 use crate::result::AnyError;
@@ -236,22 +236,20 @@ impl ModuleLoader for FsModuleLoader {
 //   }
 // }
 
-// #[derive(Default)]
-// pub struct CoreModuleLoader;
-//
-// impl ModuleLoader for CoreModuleLoader {
-//   fn resolve(&self, _: Option<&str>, specifier: &str) -> anyhow::Result<ModulePath> {
-//     match CORE_MODULES().get(specifier) {
-//       Some(_) => Ok(specifier.to_string()),
-//       None => bail!(format!("Module not found \"{specifier}\"")),
-//     }
-//   }
-//   fn load(&self, specifier: &str) -> anyhow::Result<ModuleSource> {
-//     // Since any errors will be caught at the resolve stage, we can
-//     // go ahead an unwrap the value with no worries.
-//     Ok(CORE_MODULES().get(specifier).unwrap().to_string())
-//   }
-// }
+#[derive(Default)]
+pub struct CoreModuleLoader;
+
+impl ModuleLoader for CoreModuleLoader {
+  fn resolve(&self, _: Option<&str>, specifier: &str) -> anyhow::Result<ModulePath> {
+    assert!(CORE_MODULES().contains_key(specifier));
+    Ok(specifier.to_string())
+  }
+  fn load(&self, specifier: &str) -> anyhow::Result<ModuleSource> {
+    // Since any errors will be caught at the resolve stage, we can
+    // go ahead an unwrap the value with no worries.
+    Ok(CORE_MODULES().get(specifier).unwrap().to_string())
+  }
+}
 
 #[cfg(test)]
 mod tests {
