+++ src/js/module.rs
@@ -13,7 +13,7 @@ use tracing::debug;
 // use url::Url;
 
 use crate::js::constant::{URL_REGEX, WINDOWS_REGEX};
-use crate::js::loader::{FsModuleLoader, ModuleLoader};
+use crate::js::loader::{CoreModuleLoader, FsModuleLoader, ModuleLoader};
 use crate::js::msg::JsRuntimeToEventLoopMessage;
 use crate::js::JsRuntime;
 use crate::result::{AnyError, VoidResult};
@@ -42,33 +42,34 @@ pub fn create_origin<'s>(
   )
 }
 
-// #[allow(non_snake_case)]
-// pub fn CORE_MODULES() -> &'static HashMap<&'static str, &'static str> {
-//   static VALUE: OnceLock<HashMap<&'static str, &'static str>> = OnceLock::new();
-//   VALUE.get_or_init(|| {
-//     let modules = vec![
-//       ("console", include_str!("./js/console.js")),
-//       ("events", include_str!("./js/events.js")),
-//       ("process", include_str!("./js/process.js")),
-//       ("timers", include_str!("./js/timers.js")),
-//       ("assert", include_str!("./js/assert.js")),
-//       ("util", include_str!("./js/util.js")),
-//       ("fs", include_str!("./module/fs.js")),
-//       ("perf_hooks", include_str!("./js/perf-hooks.js")),
-//       ("colors", include_str!("./js/colors.js")),
-//       ("dns", include_str!("./js/dns.js")),
-//       ("net", include_str!("./js/net.js")),
-//       ("test", include_str!("./js/test.js")),
-//       ("stream", include_str!("./js/stream.js")),
-//       ("http", include_str!("./js/http.js")),
-//       ("@web/abort", include_str!("./js/abort-controller.js")),
-//       ("@web/text_encoding", include_str!("./js/text-encoding.js")),
-//       ("@web/clone", include_str!("./js/structured-clone.js")),
-//       ("@web/fetch", include_str!("./js/fetch.js")),
-//     ];
-//     HashMap::from_iter(modules.into_iter())
-//   })
-// }
+#[allow(non_snake_case)]
+pub fn CORE_MODULES() -> &'static HashMap<&'static str, &'static str> {
+  static VALUE: OnceLock<HashMap<&'static str, &'static str>> = OnceLock::new();
+  VALUE.get_or_init(|| {
+    let modules = vec![
+      ("rsvim:ext/infra", include_str!("./runtime/00__infra.js")),
+      // ("console", include_str!("./js/console.js")),
+      // ("events", include_str!("./js/events.js")),
+      // ("process", include_str!("./js/process.js")),
+      // ("timers", include_str!("./js/timers.js")),
+      // ("assert", include_str!("./js/assert.js")),
+      // ("util", include_str!("./js/util.js")),
+      // ("fs", include_str!("./module/fs.js")),
+      // ("perf_hooks", include_str!("./js/perf-hooks.js")),
+      // ("colors", include_str!("./js/colors.js")),
+      // ("dns", include_str!("./js/dns.js")),
+      // ("net", include_str!("./js/net.js")),
+      // ("test", include_str!("./js/test.js")),
+      // ("stream", include_str!("./js/stream.js")),
+      // ("http", include_str!("./js/http.js")),
+      // ("@web/abort", include_str!("./js/abort-controller.js")),
+      // ("@web/text_encoding", include_str!("./js/text-encoding.js")),
+      // ("@web/clone", include_str!("./js/structured-clone.js")),
+      // ("@web/fetch", include_str!("./js/fetch.js")),
+    ];
+    HashMap::from_iter(modules)
+  })
+}
 
 /// Module path on local file system.
 pub type ModulePath = String;
@@ -496,16 +497,13 @@ pub fn resolve_import(
     None => specifier.into(),
   };
 
-  // // Look the params and choose a loader, then resolve module.
-  // let is_core_module_import = CORE_MODULES().contains_key(specifier.as_str());
-  // if is_core_module_import && !ignore_core_modules {
-  //   CoreModuleLoader {}.resolve(base, &specifier)
-  // } else {
-  //   FsModuleLoader {}.resolve(base, &specifier)
-  // }
-
-  // We don't actually have core modules
-  FsModuleLoader {}.resolve(base, &specifier)
+  // Look the params and choose a loader, then resolve module.
+  let is_core_module_import = CORE_MODULES().contains_key(specifier.as_str());
+  if is_core_module_import && !ignore_core_modules {
+    CoreModuleLoader {}.resolve(base, &specifier)
+  } else {
+    FsModuleLoader {}.resolve(base, &specifier)
+  }
 }
 
 /// Loads an import using the appropriate loader.
@@ -525,14 +523,15 @@ pub fn load_import(specifier: &str, skip_cache: bool) -> anyhow::Result<ModuleSo
   // // Load module.
   // loader.load(specifier)
 
-  // if CORE_MODULES().contains_key(specifier) {
-  //   CoreModuleLoader {}.load(specifier)
-  // } else {
-  //   FsModuleLoader {}.load(specifier)
-  // }
+  let is_core_module_import = CORE_MODULES().contains_key(specifier);
+  if is_core_module_import {
+    CoreModuleLoader {}.load(specifier)
+  } else {
+    FsModuleLoader {}.load(specifier)
+  }
 
-  // We don't actually have core modules
-  FsModuleLoader {}.load(specifier)
+  // // We don't actually have core modules
+  // FsModuleLoader {}.load(specifier)
 }
 
 pub async fn load_import_async(specifier: &str, skip_cache: bool) -> anyhow::Result<ModuleSource> {
