+++ src/js/runtime/10__global.d.ts
