+++ src/js/runtime/10__global.ts
@@ -1,5 +1,8 @@
 //! Js runtimes for global object, i.e. `globalThis`.
 
+// @ts-ignore Ignore internal import warning
+import infra from "rsvim:ext/infra";
+
 // `globalThis`
 interface GlobalThisType {
   setTimeout(
@@ -10,90 +13,6 @@ interface GlobalThisType {
 }
 
 ((globalThis) => {
-  // Utils {
-  /**
-   * @param {*} value
-   * @returns {string}
-   */
-  function stringify(value: any): string {
-    if (typeof value === "string") {
-      return `string["${value}"]`;
-    }
-
-    if (typeof value === "number") {
-      if (Number.isInteger(value)) {
-        return `int[${value}]`;
-      }
-
-      return `float[${value}]`;
-    }
-
-    if (typeof value === "boolean") {
-      return `boolean[${value ? "true" : "false"}]`;
-    }
-
-    if (typeof value === "function") {
-      return `function[${value.toString()}]`;
-    }
-
-    if (typeof value === "object") {
-      if (Array.isArray(value)) {
-        return `array[length: ${value.length}]`;
-      }
-
-      if (value instanceof Map) {
-        return `Map[size: ${value.size}]`;
-      }
-
-      if (value instanceof WeakMap) {
-        return `WeakMap[]`;
-      }
-
-      if (value instanceof Set) {
-        return `Set[size: ${value.size}]`;
-      }
-
-      if (value instanceof WeakSet) {
-        return `WeakSet[]`;
-      }
-
-      if (value instanceof String) {
-        return `String["${value}"]`;
-      }
-
-      if (value instanceof Number) {
-        let source = value.valueOf();
-
-        if (Number.isInteger(source)) {
-          return `Number:int[${source}]`;
-        }
-
-        return `Number:float[${source}]`;
-      }
-
-      if (value instanceof Boolean) {
-        return `Boolean[${value.valueOf() ? "true" : "false"}]`;
-      }
-
-      if (value instanceof Date) {
-        return `Date["${value.toUTCString()}"]`;
-      }
-
-      if (value instanceof RegExp) {
-        return `RegExp[${value.toString()}]`;
-      }
-
-      return `object[${JSON.stringify(value)}]`;
-    }
-
-    if (typeof value === "undefined") {
-      return "undefined";
-    }
-
-    throw new Error(`Unhandled type ${typeof value}`);
-  }
-  // Utils }
-
   // Timer API {
 
   const TIMEOUT_MAX = Math.pow(2, 31) - 1;
@@ -125,7 +44,7 @@ interface GlobalThisType {
     // Check if callback is a valid function.
     if (typeof callback !== "function") {
       throw new Error(
-        `"setTimeout" callback must be function type, but found ${stringify(callback)}`,
+        `"setTimeout" callback must be function type, but found ${infra.stringify(callback)}`,
       );
     }
